refactor: Update Zotero API Key and README
This commit refactors the display of the bibliography and updates the configuration for the Zotero API key. The display of the bibliography is still a work in progress and needs further improvements. The configuration for the Zotero API key is now stored in a separate file, "params.yaml", to prevent sharing it in the repository.

BREAKING CHANGE: